+++ tests/WebPImageConverterTest.php
@@ -37,6 +37,29 @@ class WebPImageConverterTest extends TestCase {
 			]
 		);
 
+		\WP_Mock::userFunction( 'wp_parse_args' )
+			->once()
+			->with(
+				[],
+				[
+					'quality'     => 20,
+					'max-quality' => 100,
+					'converter'   => 'gd',
+				]
+			)
+			->andReturn(
+				[
+					'quality'     => 20,
+					'max-quality' => 100,
+					'converter'   => 'gd',
+				]
+			);
+
+		\WP_Mock::userFunction( 'get_option' )
+			->once()
+			->with( 'webp_img_converter', [] )
+			->andReturn( [] );
+
 		$options = $converter->get_options();
 
 		$this->assertSame(
@@ -50,21 +73,52 @@ class WebPImageConverterTest extends TestCase {
 		$this->assertConditionsMet();
 	}
 
-	public function test_get_options_returns_modified_settings() {
+	public function test_get_options_returns_filter_settings() {
 		$converter = Mockery::mock( WebPImageConverter::class )->makePartial();
 		$converter->shouldAllowMockingProtectedMethods();
 
 		\WP_Mock::onFilter( 'webp_img_options' )
 			->with(
 				[
-					'quality'     => 20,
+					'quality'     => 75,
 					'max-quality' => 100,
-					'converter'   => 'gd',
+					'converter'   => 'imagick',
 				]
 			)
 			->reply(
 				[
 					'quality'   => 50,
+					'converter' => 'imagick',
+				]
+			);
+
+		\WP_Mock::userFunction( 'wp_parse_args' )
+			->once()
+			->with(
+				[
+					'quality'   => 75,
+					'converter' => 'imagick',
+				],
+				[
+					'quality'     => 20,
+					'max-quality' => 100,
+					'converter'   => 'gd',
+				]
+			)
+			->andReturn(
+				[
+					'quality'     => 75,
+					'max-quality' => 100,
+					'converter'   => 'imagick',
+				]
+			);
+
+		\WP_Mock::userFunction( 'get_option' )
+			->once()
+			->with( 'webp_img_converter', [] )
+			->andReturn(
+				[
+					'quality'   => 75,
 					'converter' => 'imagick',
 				]
 			);
@@ -76,6 +130,63 @@ class WebPImageConverterTest extends TestCase {
 			[
 				'quality'   => 50,
 				'converter' => 'imagick',
+			]
+		);
+		$this->assertConditionsMet();
+	}
+
+	public function test_get_options_returns_plugin_settings() {
+		$converter = Mockery::mock( WebPImageConverter::class )->makePartial();
+		$converter->shouldAllowMockingProtectedMethods();
+
+		\WP_Mock::expectFilter(
+			'webp_img_options',
+			[
+				'quality'     => 66,
+				'max-quality' => 100,
+				'converter'   => 'cwebp',
+			]
+		);
+
+		\WP_Mock::userFunction( 'wp_parse_args' )
+			->once()
+			->with(
+				[
+					'quality'   => 66,
+					'converter' => 'cwebp',
+				],
+				[
+					'quality'     => 20,
+					'max-quality' => 100,
+					'converter'   => 'gd',
+				]
+			)
+			->andReturn(
+				[
+					'quality'     => 66,
+					'max-quality' => 100,
+					'converter'   => 'cwebp',
+				]
+			);
+
+		\WP_Mock::userFunction( 'get_option' )
+			->once()
+			->with( 'webp_img_converter', [] )
+			->andReturn(
+				[
+					'quality'   => 66,
+					'converter' => 'cwebp',
+				]
+			);
+
+		$options = $converter->get_options();
+
+		$this->assertSame(
+			$options,
+			[
+				'quality'     => 66,
+				'max-quality' => 100,
+				'converter'   => 'cwebp',
 			]
 		);
 		$this->assertConditionsMet();